+++ wedding/src/server/createRouter.ts
@@ -1,0 +1,3 @@
+import * as trpc from '@trpc/server';
+
+export const createRouter = () => trpc.router();
